adding auto completion and suggestions on files and folders
added suggestion and autocompletion for files and folders.

start typing a folder or file and it will show you the rest of the
characters of the closest matching file or dir in grayed out text.

hit ENTER to accept the current suggestion, or if multiple folders are
found that starts with your input, it will enumurate the files and show
you how many more suggestions it has, hit TAB to cycle through these
suggestions and hit enter to accept current suggestion