fixed cursor jumping and redrawing when using autocompletion and suggestions
fixed a bug where every time the user hit TAB to cycle thorugh
suggestions the cursor would jump to the start of the command, then to
the end of the suggestion count and then finally to the endof the
suggestion.

now the cursor will always draw at the end of the suggestion and before
the suggestion count. making for a smoother experience

fixed drawing text

everytime the user would cycle through suggestions the whole string
inputted to the terminal would redraw for a split second and it looking
a little jarring. fixed this so whatever text is already written wont be
redrawn, and only redraws the suggestion. Making for a super smooth
experience